services/chatterbox/pyproject.toml
[project]
name = "chatterbox-service"
version = "0.1.0"
description = "OpenAI-compatible TTS API backed by Chatterbox-Turbo"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.7",
    "chatterbox-tts>=0.1.6",
    "pydub>=0.25",
    # pin onnx to avoid ml-dtypes incompatibility (onnx>=1.17 needs
    # ml-dtypes>=0.5 which conflicts with chatterbox's numpy<1.26 pin)
    "onnx<1.17",
    # perth (audio watermarking, dep of chatterbox) uses pkg_resources
    # which was removed in setuptools>=81
    "setuptools<81",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11",
    "pytest>=8",
    "httpx>=0.28",
    "pytest-asyncio>=0.25",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chatterbox_service"]

[tool.ruff]
target-version = "py311"
line-length = 99
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["chatterbox_service"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
